Feat: Implement storage swap functionality
Adds the ability to swap the source and destination storages directly within the Storage component view.

This change makes the `sourceStorageId` and `destinationStorageId` reactive using Vue `ref`. A new `swapStorage` function is introduced to exchange the values of these refs.

The `useStorageData` composable is updated to watch the passed-in `sourceStorageId` and `destinationStorageId` refs. When these IDs change, the composable automatically re-fetches the storage names, article groups (for the source), and article lists for both storages.

The swap button in the UI now triggers the `swapStorage` function via an `@click` handler, providing a smoother user experience compared to navigating to a new URL. Furthermore this way the URL does not change and no new entry is pushed onto the history, which would prevent the close button to work-

Files changed:
- `client/src/components/Storage.vue`: Made storage IDs reactive, added swap function, updated swap button.
- `client/src/composables/useStorageData.js`: Added watchers for storage ID refs to enable automatic data re-fetching on swap.