feat: Prevent storage swap when not possible
In certain modes, such as "request" mode, the source storage can be null. This change prevents the user from attempting to swap the source and destination storages when one of them is not defined.

A check has been added to the `swapStorage` function in `Storage.vue`. If a swap is attempted with a null storage, an informational modal is now displayed to the user.

New i18n keys for `storage_swap_not_possible` have been added to support this new message.